Setup compensation feature added
Added functionality impedance spectra to correct short open load
calibration data. With this data the individual wiring of the
measurement setup can be compensated.

Black formatter in the notebooks unfortunately removes the indentations,
so the exported code no longer works. This can only be used for copy and
paste without indentations.